remove unecessary exports and use pdfinverter package
