+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bytes"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -10,19 +11,11 @@ import (
 	"path/filepath"
 
 	"github.com/google/uuid"
+	"github.com/rootVIII/pdfinverter/pdfinverter"
 )
 
-// AppInitializer provides an interface to the CLI and GUI types.
-type AppInitializer interface {
-	ImageRoutine(imgName string, fin chan<- struct{})
-	ExtractImage()
-	IterImage(imgName string)
-	WritePDF()
-	RunApp()
-}
-
-// RunCLI is the entry point to the cmd-line version.
-func RunCLI(tmpDir string, pngtopdf string) {
+// runCLI is the entry point to the cmd-line version.
+func runCLI(tmpDir string, pngtopdf string) {
 	inputFile := flag.String("i", "", "input file path")
 	outputFile := flag.String("o", "", "output file path")
 	flag.Parse()
@@ -31,32 +24,58 @@ func RunCLI(tmpDir string, pngtopdf string) {
 	} else if _, err := os.Stat(*inputFile); err != nil {
 		fmt.Println("invalid file path provided for -i <input>")
 	} else {
-		var cliInit AppInitializer
-		cliInit = &CLI{
-			PDFInverter: PDFInverter{
+		var cliInit pdfinverter.PDFInverter
+		cliInit = &pdfinverter.CLI{
+			App: pdfinverter.App{
 				TmpDir:     tmpDir,
 				PDFIn:      *inputFile,
 				PDFOut:     *outputFile,
 				PyPNGToPDF: pngtopdf,
-				ImgCount:   0,
 			},
 		}
 		cliInit.RunApp()
 	}
 }
 
-// RunGUI runs the program with a QT front-end..
-func RunGUI(tmpDir string, pngtopdf string) {
+// runGUI runs the program with a QT front-end..
+func runGUI(tmpDir string, pngtopdf string) {
 
-	var guiInit AppInitializer
-	guiInit = &GUI{
-		PDFInverter: PDFInverter{
+	var guiInit pdfinverter.PDFInverter
+	guiInit = &pdfinverter.GUI{
+		App: pdfinverter.App{
 			TmpDir:     tmpDir,
 			PyPNGToPDF: pngtopdf,
-			ImgCount:   0,
 		},
 	}
 	guiInit.RunApp()
+}
+
+// getPDFConv returns Python code used as a utility shell script.
+func getPDFConv() []byte {
+	script := []byte(`
+import Quartz as Quartz
+from CoreFoundation import NSImage
+from os.path import realpath, basename
+from sys import argv
+
+
+def png_to_pdf(args):
+    image = NSImage.alloc().initWithContentsOfFile_(args[0])
+    page_init = Quartz.PDFPage.alloc().initWithImage_(image)
+    pdf = Quartz.PDFDocument.alloc().initWithData_(page_init.dataRepresentation())
+
+    for index, file_path in enumerate(args[1:]):
+        image = NSImage.alloc().initWithContentsOfFile_(file_path)
+        page_init = Quartz.PDFPage.alloc().initWithImage_(image)
+        pdf.insertPage_atIndex_(page_init, index + 1)
+
+    pdf.writeToFile_(realpath(__file__)[:-len(basename(__file__))] + 'aggr.pdf')
+
+
+if __name__ == '__main__':
+	png_to_pdf(argv[1:])
+`)
+	return bytes.ReplaceAll(script, []byte{0x09}, []byte{0x20, 0x20, 0x20, 0x20})
 }
 
 func main() {
@@ -82,12 +101,16 @@ func main() {
 	}
 
 	pngtopdfTMP := filepath.Join(tmpdir, randFileName.String())
-	tmpdir += "/"
-	WriteText(pngtopdfTMP, GetPDFConv())
 
+	err = ioutil.WriteFile(pngtopdfTMP, getPDFConv(), 0700)
+	if err != nil {
+		panic(err)
+	}
+
+	tmpdir += "/"
 	if len(os.Args) > 1 {
-		RunCLI(tmpdir, pngtopdfTMP)
+		runCLI(tmpdir, pngtopdfTMP)
 	} else {
-		RunGUI(tmpdir, pngtopdfTMP)
+		runGUI(tmpdir, pngtopdfTMP)
 	}
 }
